// packages/butter-widget/src/utils/data-types.ts
export type HomepageVolumeChartData = {
  volume: number;
  chainName: string;
  color: string;
  sourceChainId: number;
  chainImg: string;
};

export type HomepageTransactionNumberData = {
  allCount: number;
  nativeCount: number;
  crossCount: number;
};

export type HomeAccountChartDataModel = {
  name: string;
  size: number;
  value: number;
};

export type HomepageChartsData = {
  crossDay30Account: any;
  day30AllVolume: string;
  day1AllVolume: string;
  day7AllVolume: string;
  allVolume: string;
  assetCharData: {
    [key: string]: number;
  };
  volumeCharData: HomepageVolumeChartData[];
  volumeCharDataDay30: HomepageVolumeChartData[];
  volumeCharDataDay7: HomepageVolumeChartData[];
  volumeCharDataDay1: HomepageVolumeChartData[];
  allTransactionCount: HomepageTransactionNumberData;
  crossChainTransactions: {
    [key: string]: number;
  };
  day1CrossChainTransactions: {
    [key: string]: number;
  };
  day7CrossChainTransactions: {
    [key: string]: number;
  };
  day30CrossChainTransactions: {
    [key: string]: number;
  };
  day1TransactionCount: HomepageTransactionNumberData;
  day7TransactionCount: HomepageTransactionNumberData;
  day30TransactionCount: HomepageTransactionNumberData;
  transactionCharData: {
    [key: string]: number;
  };
  allTransactionCharData: {
    [key: string]: number;
  };
  nativeTransactionCharData: {
    [key: string]: number;
  };
  crossChainTransactionCharData: {
    [key: string]: number;
  };
  allAccount: number;
  day1Account: number;
  day7Account: number;
  day30Account: number;
  accountCharData: {
    [key: string]: number;
  };
  contractCharData: {
    [key: string]: number;
  };
  allContractCharData: {
    [key: string]: number;
  };
  nativeContractCharData: {
    [key: string]: number;
  };
  crossContractCharData: {
    [key: string]: number;
  };

  contractCount: string;
  accountDay1DataArr: HomeAccountChartDataModel[];
  accountDay7DataArr: HomeAccountChartDataModel[];
  accountDay30DataArr: HomeAccountChartDataModel[];
  accountAllDataArr: HomeAccountChartDataModel[];
  allContractCount: HomepageTransactionNumberData;
  day1ContractCount: HomepageTransactionNumberData;
  day7ContractCount: HomepageTransactionNumberData;
  day30ContractCount: HomepageTransactionNumberData;
};
// ...
export type ChartsData = {
  labels: string[];
  values: number[];
};